Homework done

+++ src/com/company/Employee.java
@@ -3,11 +3,11 @@ package com.company;
 public class Employee {
     String firstName;
     String lastName;
-    int salary;
+    double salary;
     int id;
     static int count=0;
 
-    Employee(String firstName,String lastName,int salary){
+    Employee(String firstName,String lastName,double salary){
         this.firstName=firstName;
         this.lastName=lastName;
         this.salary=salary;
